[Examples] Add `BasicAssetLibrary` example manager
In order to develop simple end-to-end tests for OpenAssetIO based
workflows, there is often a need for a simple, controllable,
deterministic asset manager. The Basic Asset Library [BAL] is just this.

Presently, it simply loads some data from a json file, and feeds that
back through the API as requested. Only `resolveEntityReference` is
supported right now. Over time, we plan to add simple registration
and version management and attribute support.

We chose to use a static JSON file as it makes creating an instance of
the manager in a known state trivial, and has minimal runtime
dependencies.

See `README.md` for more information.

BAL is not in competition with, or replace the Sample Asset Manager
example - which will, over time, support the full range of API
functionality. BAL remains simple, the runtime dependencies of SAM will
certainly grow beyond what is desirable for a simple test harness puppet.

Part of #215.

Signed-off-by: Tom Cowland <tom@foundry.com>