[Core] Switch to unified entity data model
This migrates registration and queries to use the new unified entity
data model. Instead of a primary string + attributes, there is a single
API call to store data (register) and one to query (resolve).

The overarching benefit to this is that it no longer strongly associates
the API with a "string to string" model, where additional properties are
an afterthought. This makes the representation of file based assets with
additional data, and non-file based assets seamless and far more
intuitive.

Resolution is parameterised by a trait set, that defines which data
is being requested by the host. This mechanism is described in more
detail in decisions/DR008-Unify-the-entity-data-model.md.

Signed-off-by: Tom Cowland <tom@foundry.com>