[Core] Remove pure-python `Specification` implementation
The C++ implementation outlined in #245 makes this historical python
Specification system redundant. This removes the old implementation,
and migrates the minimal existing use (mostly in the test harness)
over to the new mechanism.

This commit temporarily hoists the C++ implementation up to the
top-level name space, pending imminent work to move it there directly.

Closes #342.

Signed-off-by: Tom Cowland <tom@foundry.com>