[Core] Split data container from `Specification` into `TraitsData`
The existing approach of Specification being a base class for custom
views _and_ the container holding the data the views operate on created
assorted problems:

- It has too many responsibilities.

- It was expected that default constructed Specification derived classes
  imbued the specification with their defined trait set. Yet, at the same
  time, passing a base class specification into derived one's constructor
  was meant to transparently 'wrap' this data to allow introspection without
  imbuing any new specifications. So the observed behaviour is divergent,
  and counterintuitive.

- The above also implied that it was down-casting the specification
  which, is not the case as the data from resolve etc. is never going
  to be an instance of the more derived class.

- Passing the specification instance into register opened up the
  possibility of someone wrapping resolved data in a view, modifying it,
  then passing this back to register. If the resolved data was missing
  any of the traits of that specification, they would never be imbued as
  they are only added to default constructed versions.

This commit separates out the data from the view.

 - TraitsData is the transport-layer container for a trait set and any
   trait property values.

 - Specification becomes the base class for strongly-typed views on
   an a TraitsData instance.

This resolves most the majority of the above concerns, and better
reenforces the requirement for the low-level transport of API data to
be simple and generic.

We spent a suitable amount of time going around in circles, and naming
wise, TraitsData won out over TraitData, TraitsInstance,
SpecificationData and values as it best defines its applicability is
limited to traits, and not all traits have values.

Closes #348.

Signed-off-by: Tom Cowland <tom@foundry.com>