[Examples] Support custom `managementPolicy` responses in BAL
Adds the ability to customize the `managementPolicy` response returned
by BAL. Amended defaults, or trait set specific policies can be supplied
using new keys in the library JSON file.

We decided to keep-it-simple for now, and just have the JSON hold the
raw policy value. Its a little opaque, but we can revisit this when we
have more use cases.

NB: This also follows the "BAL style" of "slightly rough and ready
programming".

Closes #459

Signed-off-by: Tom Cowland <tom@foundry.com>