[Core] Return traits from `managementPolicy`
Part of #492. Redesign `managementPolicy` to return `TraitsData` rather
than a bitfield.

Add a `ManagedTrait` to replace `kIgnored` (by absence of the trait),
`kMananged` (by presence of the trait) and `kExclusive` (via a trait
property).

Since we are moving to traits for presenting `managementPolicy`
policies, remove the `kWillManagePath` flag and update the docs to
use a `WillManagePath` trait instead.

Similarly, remove the `kWantsThumbnail` flag and update the docs to
use a `WantsThumbnail` trait instead. Don't add any code to support
this, since thumbnails are industry-specific so any traits should be
defined in an industry-specific repository (e.g.
OpenAssetIO-MediaCreation). Hence the docs just refer to an imagined
trait, in much the same way as various other examples.

Update BAL test manager plugin to return `TraitsData` constructed from
its JSON library for its `managementPolicy` response.

Signed-off-by: David Feltell <david.feltell@foundry.com>