[Core] Migrate `HostSession` logging to C++ and refactor
Migrates the remaining `HostSession` functionality to C++, facilitating
the removal of the Python wrapper.

As part of this work, we elected to remove the log fire-break.
Primarily on the grounds of "does it really add that much, for the
maintenance overhead". We can still abstract out changes in the host
facing API through an implementation of LoggerInterface that serves
as an adapter.

Closes #531.

Signed-off-by: Tom Cowland <tom@foundry.com>