[Core] Rename `isEntityReference` to `isEntityReferenceString`
Renames to avoid ambiguity when also working with the new
`EntityReference` type - this method does not tell you if something
is an instance of one of those types.

Opted to update the recent DRs around entity reference workflows as
it was those that highlighted the ambiguity, and this hopefully
clarifies the workflow options they illustrate.

Part of #549.

Signed-off-by: Tom Cowland <tom@foundry.com>