[Core] Add `kEntityAccessError` to `BatchElementError`
Allows us to specifically identify cases where the Context's access is
the cause of the failure.

Generally this is read .vs. write, and could be quite common. So
understanding this specifically is important.

This consciously renumbers the enums as we're already doing that in the
next release, and this error is applicable to many methods, where as the
resolution error is more specific.

Part of #587.

Signed-off-by: Tom Cowland <tom@foundry.com>