[Core] Add `kMalformedEntityReference` code to `BatchElementError`
Hopefully this finally aligns our handling of how to define the state
of an entity reference. We should now have these states:

 - Unknown - a string that has not been tested yet.
 - Valid - a string that is recognised by a manager (eg: correct
   URI scheme).
 - Invalid - a string that is not recognised by a manager.
 - Malformed - A valid reference that is malformed dor the context
   of use.

`createEntityReferenceIfValid` and `kInvalidEntityReference` now use
"valid" in the context of the above.

The new `kMalformedEntityReference`/`MalformedEntityReference` exception
and the `apiComplianceSuite` tests now use "malformed" in the context
of the above.

I didn't document the new Exception in the release notes as we should
be phasing out the exception as we migrate methods to callbacks, and
update any non-callback versions into hold a `BatchErrorCode` rather
than an `Exception` in the result variant.

Part of #587.

Signed-off-by: Tom Cowland <tom@foundry.com>