[Examples] Add basic publishing support to BAL
Adds basic publishing support to BAL. Limited to the blanket
creation/update of entities, with no validation of whether the new
trait set matches any data previously held for the entity.

All updates to the library are currently in-memory only and can't be
persisted to disk. Internally, new versions are added by the publishing
operation, though we don't expose version functionality yet. This should
allow this to be an easy addition in the future.

Though limited, this support serves good purpose for validating
simple publishing workflows around the updated concepts for API
invocation and error handling (illustrated by the preceding commits,
the requirements for which were surfaced by this work, see #591, #594,
#595).

Part of #585.

Signed-off-by: Tom Cowland <tom@foundry.com>