[Plugin] Expose `openassetio.manager_plugin` entry point
When installed via pip or similar, the BAL package now exposes an
`openassetio.manager_plugin` entry point, allowing it to be
discovered/loaded without needing to extend `$OPENASSETIO_PLUGIN_PATH`.

This paves the way to releasing BAL as a packaged to PyPI to futher
simplify installation and use in downstream projects.

Closes #9

Signed-off-by: Tom Cowland <tom@foundry.com>