[Core] Remove hardcoded default managementPolicy
Closes #6. The managementPolicy is industry specific, BAL should be a
generic stub/puppet manager that can be used in any area that the main
OpenAssetIO library is employed. Consequently it should not have
industry-specific references/dependencies.

So remove the hardcoded default fallback policy and instead raise a
human-readable exception if there isn't a (default) policy provided in
the (industry-specific) JSON library.

Signed-off-by: David Feltell <david.feltell@foundry.com>