[Core] Use `dataclass` instead of `namedtuple`
We had originally decided to make this change to avoid potential
foot-guns from to default list arguments. However, when it came to it,
we weren't actually ever using the defaults.

Seems like a worthwhile change regardless as it is far more readable,
and affords us the ability to include type hints.

Closes #20.

Signed-off-by: Tom Cowland <tom@foundry.com>