[Core] Improve handling of empty library path strings
Fallback to BAL_LIBRARY_PATH was inadvertently broken when we added the
settings mechanism. This is due to the fact we never considered a
setting set to an empty string as being "unset".

This improves behaviour such that if the `library_path` setting is set,
but to an emtpy string, we still fall back on the env var. Only then, if
we have an empty string/None, then we grumble.

This rejigs the logic a little in init to also ensure we don't update
the settings in the manager if we raise an exception due to library path
being empty. This makes things more atomic, which is a good thing.

NB. This bumps the required OpenAssetIO version to alpha-9 for tests.

Closes #26

Signed-off-by: Tom Cowland <tom@foundry.com>