[Core] Fix entityExists for malformed references
This used to return an exception, which is not part of the API contract.
This method is soon to migrate to callbacks, at which point this will be
an error callback invocation.

I haven't added any test for this as apiComplianceSuite should be
testing this case when we've moved over to callbacks.

Signed-off-by: Tom Cowland <tom@foundry.com>