[Core] Add entityTraits
Closes #89. The `entityTraits` API method is a new required method.

For reading, return the traits of the entity, plus the implicit
VersionTrait. For write, return the traits of the entity, but don't
include the VersionTrait (since it cannot be written directly).

The implementation assumes new entities have no restriction on the
traits that can be published to them; whereas new versions of existing
entities are restricted to the same trait set as previous versions.
This latter case is not actually enforced during `preflight`/`register`.
See #92.

Signed-off-by: David Feltell <david.feltell@foundry.com>