[Core] Fix resolve when trait has no properties
Discovered whilst updating OpenAssetIO/OpenAssetIO-MediaCreation#69
after OpenAssetIO/OpenAssetIO#1217. The API docs for `resolve` say

> Only traits that are applicable to each entity, and for which the
  manager has data, will be imbued in the result.

However, if an entity in the JSON library has a trait as part of its
trait set, but that trait has no properties (either because that trait
has no associated properties, or because it's not set in the JSON), then
it was imbued in the `TraitsData` regardless.

Signed-off-by: David Feltell <david.feltell@foundry.com>