[Core] Support entity access override
Part of #90. Support an `"overrideByAccess"` dict for each entity in the
JSON database (mirroring the `"overrideByTraitSet"` dict for
`managementPolicy`), where the key is a stringified access mode and the
value is an alternative version dict to return for that access mode.

This then allows customising the returned set of traits (and their
property data) per access mode, supporting both `resolve` and
`entityTraits`.

An override of `null` indicates the entity doesn't exist for that
access mode, i.e. `kEntityResolutionError`.

An override of an empty dict (or rather, a dict without `"traits"`)
indicates the entity is inaccessible for that access mode, i.e.
`kEntityAccessError`.

This mechanism thus replaces the buggy `supported_access_modes` list,
which only supported `resolve`, and would reset to read-only whenever an
entity was published.

Access mode overrides for relationship queries are not yet supported
and will be the subject of future work.

Signed-off-by: David Feltell <david.feltell@foundry.com>