[Core] Validate publish against policy
Part of #90. Validate `preflight` and `register` against the management
policy for the trait set.

This then allows specific (or all) trait sets to be allowed or
disallowed for publishing.

A change in behaviour is that some non-empty policy _must_ be specified
in the JSON database for `kWrite` in order for publishing to succeed.

Signed-off-by: David Feltell <david.feltell@foundry.com>