[Core] Rename "exceptions"->"overrideByTraitSet"
Part of #90. The response from `managementPolicy` can be customised per
trait set through an option in the JSON database. This option was called
`"exceptions"`, which is not consistent with the naming of the new
entity traits feature `"overrideByAccess"`. We decided that "override"
is less of an overloaded term than "exception", and that we should be
more informative by being specific about the key we're overriding by.

So rename the JSON key for specifying per trait set overrides of
`managementPolicy` responses from `"exceptions"` to
`"overrideByTraitSet"`.  This is then consistent with
`"overrideByAccess"` and more informative than before.

Signed-off-by: David Feltell <david.feltell@foundry.com>