[Core] Don't %-decode file URLs from DB
Closes #109. We should not be percent-decoding URLs fetched from the
JSON DB before passing to the host, since other functions that consume
the returned URL will most likely expect the URL to be percent-encoded.

In order to allow for upward traversals to be used in `file://` URLs in
the DB (in particular, useful for e.g.
`${bal_library_dir_url}/../path/to/file`) we needed to parse the URL to
treat as a path, resolve the upward traversals (since they are invalid
in URLs according to `FileUrlPathConverter`), then reconstruct the URL.
Unfortunately in doing this we also percent-decoded the URL and didn't
re-encode it.

Passing the decoded URL to `FileUrlPathConverter.pathFromUrl` would then
do a second round of percent-decoding, possibly decoding what should
have been left as a %.

In addition, fully converting the URL to a path and then re-encoding as
a URL causes other issues when we consider cross-platform paths. E.g. a
`file:///mnt/` URL intended for POSIX may well be read on a Windows host
(for whatever reason), and in that case we shouldn't be trying to
normalise based on Windows rules, where it will be treated as a relative
path and have the drive letter of the current working directory prefixed
to it, resulting in `file:///C:/mnt/`.

So dodge the issue of platform-specific behaviour and percent
decoding/encoding entirely, by extracting the path component and
treating as a POSIX path for the purposes of normalisation, before
reconstituting.

Signed-off-by: David Feltell <david.feltell@foundry.com>