[Core] Configurable `hasCapability` responses
Closes #84. Host applications must be able to tolerate manager plugins
that do not support the full suite of (optional) capabilities that a
manager may implement. However, there was no way for hosts to use BAL
to test their logic for dealing with different combinations of
capabilities.

In addition, configurable capabilities are required to enable e2e
testing of the capability-based routing used by the upcoming hybrid
plugin system (see OpenAssetIO/OpenAssetIO#1202).

So add a simple way to override the default set of capabilities reported
by BAL, by adding an optional "capabilities" list element to the
library, where each element is a stringified capability, as defined in
`kCapabilityNames`. Presence of a capability in this list indicates that
it is supported. If the list is not found, then the default (i.e. true)
set of capabilities is used.

Short-circuit methods where the capability is unsupported. Instead, call
the base class (which will raise an error).

This logic is similar to that used in the SimpleCppManager (see
OpenAssetIO/OpenAssetIO#1324).

Signed-off-by: David Feltell <david.feltell@foundry.com>