[Core] Allow overriding of identifier
Needed for e2e testing of OpenAssetIO/OpenAssetIO#1202. The upcoming
hybrid plugin system will work by locating multiple plugins that all
advertise the same unique identifier and combining them.

In preparation for e2e testing of this, make BAL's plugin identifier
configurable by an environment variable, so that it can be set to the
same identifier as another plugin on the system.

We cannot use other types of settings such as the JSON database, since
the plugin identifier is queried before the settings are provided.

Signed-off-by: David Feltell <david.feltell@foundry.com>