[Build] Bump OpenAssetIO to 1.0.0-beta.2.2
Discovered during OpenAssetIO/OpenAssetIO#1311.

Since #84, BAL's tests make use of the newer `getWithRelationship`
signatures and so fail if OpenAssetIO 1.0.0-beta.2.1 or earlier is
installed.

This isn't an incompatibility in BAL itself, just its tests. We could
either:
* Rewrite the tests to use the older API.
* Bump the minimum supported version of OpenAssetIO
* Ignore the failure, since it's only tests.

Ignoring the test failure is potentially dangerous, in case the expected
failure is masking a real issue.

Rewriting the tests would be the best option, but requires
development effort. Given that we're still in beta, the low-effort
choice of bumping the minimum OpenAssetIO version seems the best
pragmatic option.

Signed-off-by: David Feltell <david.feltell@foundry.com>