[Core] Support defaultEntityReference
Closes #53. Add support for the `defaultEntityReference` API method,
which provides a way for hosts to get a "starting point" for further
queries.

Add a new top-level dictionary to the JSON database `defaultEntities`,
which has a mapping per access mode of trait set to entity reference.

If no matching trait set is found in the JSON database, then respond
with `kInvalidTraitSet`. In real-world managers, an unrecognised trait
set would probably elicit a response of `None` (for "no appropriate
default"), rather than exception. But here we make a `None` response
explicit by specifying `null` in the JSON database.

Signed-off-by: David Feltell <david.feltell@foundry.com>