[Core] Add relationship access and version overrides
Part of testing TheFoundryVisionmongers/KatanaOpenAssetIO#4.

BAL would error when performing relationship queries with a non-`kRead`
access mode. However, such queries are valid for publishing workflows
and so must be supported in order to test host applications.

So add an optional `"access"` flag that can be added to relationship
definitions in the JSON database, overriding the default `kRead` access,
and limiting that relationship to specific access modes.

Hence no longer error when responding to relationship queries for
non-`kRead` access modes (the default response instead being an empty
set of entity references).

Additionally, if a version relationship trait set is specified in the
JSON database, then prefer that over the built-in versioning behaviour.
A long term goal would be to remove the built-in versioning behaviour in
favour of leveraging generic relationship queries (#88). This at least
gets us part way there to support some use cases.

Signed-off-by: David Feltell <david.feltell@foundry.com>